# web/api-docs/swagger.yaml
basePath: /
definitions:
...
  github_com_cuihe500_vaulthub_internal_database_models.SafeUserProfile:
    properties:
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: integer
      nickname:
        type: string
      phone:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  github_com_cuihe500_vaulthub_internal_service.CreateProfileRequest:
    properties:
      email:
        maxLength: 100
        type: string
      nickname:
        maxLength: 50
        type: string
      phone:
        maxLength: 20
        type: string
    required:
    - email
    - nickname
    type: object
...
  github_com_cuihe500_vaulthub_internal_service.ListProfilesResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      profiles:
        items:
          $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SafeUserProfile'
        type: array
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  github_com_cuihe500_vaulthub_internal_service.UpdateProfileRequest:
    properties:
      email:
        maxLength: 100
        type: string
      nickname:
        maxLength: 50
        type: string
      phone:
        maxLength: 20
        type: string
    type: object
...
  github_com_cuihe500_vaulthub_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      requestId:
        type: string
      timestamp:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: admin@thankseveryone.top
    name: API Support
  description: |-
    VaultHub 是一个密钥管理系统，旨在安全地存储、管理和轮换加密密钥、API 密钥及其他敏感凭证。

    ## 特性
    - 安全的密钥存储和管理
    - 密钥轮换支持
    - 细粒度的访问控制
    - 审计日志

    ## 认证
    大部分接口需要JWT Bearer Token认证。
    1. 调用 /api/v1/auth/register 注册账号
    2. 调用 /api/v1/auth/login 获取 token
    3. 在请求头中添加 Authorization: Bearer {token}

    ## 响应格式
    所有接口返回统一的响应格式，HTTP状态码均为200，错误码在响应体的code字段中：
    ```json
    {
    "code": 0,
    "message": "success",
    "data": {},
    "requestId": "xxx",
    "timestamp": 1762269490888
    }
    ```
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://github.com/cuihe500/vaulthub
  title: VaultHub API
  version: "1.0"
paths:
  /api/v1/admin/profiles:
    get:
      consumes:
      - application/json
      description: 获取用户档案列表（需要管理员权限）。不传分页参数时全量导出（最多10000条）
      parameters:
      - description: 页码（可选，不传则全量导出）
        in: query
        minimum: 1
        name: page
        type: integer
      - description: 每页数量（可选，不传则全量导出）
        in: query
        maximum: 10000
        minimum: 1
        name: page_size
        type: integer
      - description: 昵称筛选
        in: query
        name: nickname
        type: string
      - description: 邮箱筛选
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.ListProfilesResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户档案列表
      tags:
      - 用户档案
  /api/v1/admin/users/{user_id}/profile:
    get:
      consumes:
      - application/json
      description: 根据用户ID获取指定用户的档案信息（需要管理员权限）
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SafeUserProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取指定用户档案
      tags:
      - 用户档案
    put:
      consumes:
      - application/json
      description: 根据用户ID更新指定用户的档案信息（需要管理员权限）
      parameters:
      - description: 用户ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: 用户档案信息
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SafeUserProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新指定用户档案
      tags:
      - 用户档案
...
  /api/v1/profile:
    delete:
      consumes:
      - application/json
      description: 删除当前用户的档案信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除用户档案
      tags:
      - 用户档案
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的档案信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SafeUserProfile'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 获取用户档案信息
      tags:
      - 用户档案
    patch:
      consumes:
      - application/json
      description: 创建或更新当前用户的档案信息（如果不存在则创建，存在则更新）
      parameters:
      - description: 用户档案信息
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.CreateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SafeUserProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建或更新用户档案
      tags:
      - 用户档案
    post:
      consumes:
      - application/json
      description: 为当前用户创建档案信息
      parameters:
      - description: 用户档案信息
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.CreateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SafeUserProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建用户档案
      tags:
      - 用户档案
    put:
      consumes:
      - application/json
      description: 更新当前用户的档案信息
      parameters:
      - description: 用户档案信息
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_service.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/github_com_cuihe500_vaulthub_internal_database_models.SafeUserProfile'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_cuihe500_vaulthub_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新用户档案
      tags:
      - 用户档案
...
securityDefinitions:
  BearerAuth:
    description: 输入 "Bearer" 后跟一个空格和 JWT token。
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"